Custom Scopes & Component Dependencies.
how to create and use custom scopes. This is useful if we want to reuse single instances of dependencies for example in an activity or fragment, but not keep them as application-wide singletons that stay in memory for the whole lifetime of the app.
To achieve this, we have to create a custom Java Annotation which itself is annotated with @Scope (and the two meta-annotations @Documented and @Retention), and use it in conjunction with a Dagger Component that lives as long as this particular scope, for example in the lifecycle of an activity or as long as a user is logged in. If we want to create an application-wide singleton, we have to instantiate its component in the Application class and access it from everywhere else through a call to getApplication.
To connect 2 components of different scopes into one dependency graph, we can declare the outer component as a dependency of the inner component and pass it as a Builder argument. The outer component has to expose dependencies to the inner component explicitly via provision methods.